requirements.txt
# Default requirements
black
flake8
isort
loguru
mkdocs
pip
python-dotenv
tqdm
typer
-e .

# Workflow requirements
ipykernel
biopython>=1.80
numpy>=2.0.0
pydna>=5.2.0
pillow>=10.3.0
scipy>=1.14.0
pandas>=2.2.2
primer3-py>=2.0.3
teemi>=0.3.4

# App requirements
dash>=2.17.1
dash-bootstrap-components>=1.6.0